add user mapping to trip

+++ scooterMS/src/main/java/tpe/scooterMS/DTO/TripRequestDTO.java
@@ -6,15 +6,9 @@ import lombok.Data;
 @Data
 public class TripRequestDTO {
 	@NotNull(message = "idUser shouldn't be null")
-	private int idUser;
+	private long idUser;
 	@NotNull(message = "idScooter shouldn't be null")
 	private int idScooter;
 	@NotNull(message = "idOriginStop shouldn't be null")
 	private int idOriginStop;
 }
-//public class TripRequestDTO {
-//	private User user;
-//	private Scooter scooter;
-//	private Stop destinationStop;
-//	private Stop originStop;
-//}
